Re-organize projects into a single solution and fix configurations
All projects (rpkg dll, cli, and gui) have been merged into a single
solution and their configurations have been updated in order to be able
to properly compile in both Release and Debug mode and for the GUI
to use the build version of the DLL without requiring manually copying
it.

The GUI has also been updated to use the project-based NuGet package
definition schema instead of the old packages.config method.