[RPKG] New Entity Deep Search Features...
Added the ability to select a search 'category' on a per search item/term basis.

The categories include:
- ALL
- tempHash (or IOI Path of)
- tbluHash (or IOI Path of)
- rootEntity
- entities
- propertyOverrides
- overrideDeletes
- pinConnectionOverrides
- pinConnectionOverrideDeletes
- externalScenes
- subType
- extraFactoryDependencies
- extraBlueprintDependencies
- comments

In addition when you use more than one search item/term they operate in conjunction in an 'AND' fashion, i.e., they must all have been found in a single entity (TEMP/TBLU) for the results to be displayed for that entity.

Fixed an issue with the Entity Deep Search JSON output function (shown after a successful search) failing on non-objects.