[RPKG] Add H1 Alphas, Beta, And Steam Support
Added preliminary support for Hitman 1 Alphas, Beta, and Steam, which now allows you to load, search, extract, etc from these older RPKG variants.

NOTE: If you want to load the RPKGs from the 'Hitman Closed Alpha' then you must change their extensions to '.rpkg' from '.pc_rpkg'. It would have not been worth it to allow for the loading/searching/etc of '.pc_rpkg' files RPKG-wide, especially when it only applies to one of the Alphas.

This does not necessarily mean that things like the older versions of models, textures, sounds are now supported, hence the 'preliminary' :)

Fixed issue in the GUI where in a recent commit clicking on the 4 char resource types disabled the Details window in RPKG.