[RPKG] Fixed Hash / IOI String Conversion Bug
Functions across RPKG that convert input JSON files to other formats used to merely check for the existence of '[' in a given input hash / IOI string value to determine if the input should be IOI(MD5)-hashed or not. If not IOI(MD5)-hashed then it was 'assumed' that the value was a valid hash. If for instance a user put in "testing" as the input value in the JSON then as AnthonyFuller points out below, since there is no '[' in that string and since there are no numbers in that string then std::strtoull() would return a value of 0.

So, now per this fix, it first checks the input string and determines if it is a viable hash value, i.e., in the format 00XXXXXXXXXXXXXX where the X's are a char in the range 0-9 or A-F, and if the input string is not a valid hash then the string is MD5-hashed and that return value is then used.

So for instance, before the fix "testing" would result in 0000000000000000, and now it results in 002B1FCA515949E5, which is the IOI(MD5)-hash of "testing".

Fixed this bug reported by AnthonyFuller:
"The ASVA -> JSON function checks to see if [ is in the string to decide if it should hash the string, I feel this isn't really the best option for it as if someone just puts a string in it there e.g. my_new_file, it won't be hashed. Which of course, isn't best practice from the user, but it's artificially limiting. The ASVA -> JSON function, if the "ASVA" value is not a valid number or contains [ will give a hash value of 0, producing this:
0000000000000000.ASVA
0000000000000000.ASVA.meta
Which is far from ideal, this should be revised."