Bone mismatch fix
The GLTF plugin in Blender appends a 'neutral_bone' sometimes when exporting to GLB and this can cause a bone 'mismatch' when trying to rebuild weighted PRIMs in RPKG.

This 'fix' just ignores the 'neutral_bone's when rebuilding the weighted PRIMs instead of giving a 'mismatch' error and stopping the rebuilding.

In cases where there is a true bone mismatch, the offending mismatched bone is now also displayed on the error message popup.

Also noted here:
https://wiki.redmodding.org/cyberpunk-2077-modding/for-mod-creators/3d-modelling/troubleshooting-your-mesh-edits#bone-neutral_bone-not-present-in-export-rig-s-import-mesh